+++ CHANGELOG.md
@@ -9,6 +9,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+- Ray tracing support for `vkCmdTraceRaysIndirectKHR` and dynamic stack size
 - Expanded the number of functions and scopes profiled by the `profiling` crate
 - Increase `PoolInfo::DEFAULT_RESOURCE_CAPACITY` from 4 to 16 in order to prevent excess resource
   creation